Uplift of #3725 (squashed) to dev

+++ build/mac/sign_app.sh
@@ -2,8 +2,12 @@
 
 set -euo pipefail
 
+usage() {
+  echo "usage: $0 <input_dir> <output_dir> <packaging_dir> <is_development> <mac_provisioning_profile> <mac_signing_keychain> <mac_signing_identifier> <notarize> <notary_user> <notary_password>"
+}
+
 if [[ ${#} -lt "7" ]]; then
-  echo "usage: $0 <input_dir> <output_dir> <packaging_dir> <is_development> <mac_provisioning_profile> <mac_signing_keychain> <mac_signing_identifier>"
+  usage
   exit 1
 fi
 
@@ -23,6 +27,24 @@ MAC_SIGNING_IDENTIFIER="${7}"
 if [[ -z ${7} ]]; then
   echo "Error: mac_signing_identifier is empty. Cannot sign."
   exit 1
+fi
+
+if [[ ${#} -gt "7" ]]; then
+  if [[ "${8}" = "True" ]]; then
+    NOTARIZE="--notarize"
+    NOTARY_USER="${9}"
+    NOTARY_PASSWORD="${10}"
+    if [[ -n "${NOTARIZE}" ]]; then
+        if [[ ( -z "${NOTARY_USER}" ) || ( -z "${NOTARY_PASSWORD}" ) ]]; then
+            echo "Error: when <notarize> is True, both <notary_user> and <notary_password> must be provided. Cannot perform notarization."
+            usage
+            exit 1
+        fi
+    fi
+    NOTARIZE_ARGS="${NOTARIZE} --notary-user $NOTARY_USER --notary-password $NOTARY_PASSWORD"
+  else
+    NOTARIZE="False"
+  fi
 fi
 
 function check_exit() {
@@ -45,11 +67,13 @@ export MAC_PROVISIONING_PROFILE
 # Clear output directory. It seems GN auto-creates directory path to the
 # expected outputs. However, the signing script doesn't expect the path to
 # have been created and fails trying to create it again.
-echo "Cleaning $DEST_DIR ..."
-rm -rf $DEST_DIR/*
+if [ -d $DEST_DIR ]; then
+  echo "Cleaning $DEST_DIR ..."
+  rm -rf $DEST_DIR/*
+fi
 
 # Invoke python script to do the signing.
-PARAMS="--input $SOURCE_DIR --output $DEST_DIR --keychain $MAC_SIGNING_KEYCHAIN --identity $MAC_SIGNING_IDENTIFIER --disable-packaging --no-notarize"
+PARAMS="--input $SOURCE_DIR --output $DEST_DIR --keychain $MAC_SIGNING_KEYCHAIN --identity $MAC_SIGNING_IDENTIFIER --disable-packaging"
 if [[ -z "${DEVELOPMENT}" ]]; then
   # Copy mac_provisioning_profile to the packaging_dir since that's where the
   # signing scripts expects to find it.
@@ -57,4 +81,11 @@ if [[ -z "${DEVELOPMENT}" ]]; then
 else
   PARAMS="$PARAMS $DEVELOPMENT"
 fi
+
+if [[ "${NOTARIZE}" = "False" ]]; then
+  PARAMS="$PARAMS --no-notarize"
+else
+  PARAMS="$PARAMS ${NOTARIZE_ARGS}"
+fi
+
 "${PKG_DIR}/sign_chrome.py" $PARAMS
